+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 # Unreleased
 
+- Support `@alias` fragments.
+
 # 3.0.0-rc.9
 
 - Fix client extension enums mixup with `FutureAddedValue`.